fix(garmin): represent segment tolerance as float end-to-end
Address PR review: SegmentDefinition.tolerance_meters is now float to match the
stored NUMERIC match_tolerance_meters and the float tolerance used by the
matching helper, so GET /segments/{id}/efforts no longer truncates a fractional
stored tolerance to int.

+++ app/models/garmin.py
@@ -520,7 +520,7 @@ class SegmentDefinition(BaseModel):
     start_lon: float = Field(description="Segment start longitude")
     end_lat: float = Field(description="Segment end latitude")
     end_lon: float = Field(description="Segment end longitude")
-    tolerance_meters: int = Field(description="Corridor radius around the start/end points in meters")
+    tolerance_meters: float = Field(description="Corridor radius around the start/end points in meters")
 
 
 class SegmentEffort(BaseModel):
